backend/requirements.txt
#fastapi, sqlModel, WebSocket requirements
fastapi
sqlmodel
websockets
uvicorn
pydantic
firebase-admin
python-jose
passlib[bcrypt]